+++ src/updater.rs
@@ -1,2 +1,3 @@
 pub mod update_manager;
 pub mod github_api_args;
+mod self_update;